% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CEDA.R
\name{alphaBeta}
\alias{alphaBeta}
\title{Calculating a significance score of a gene based on 
the corresponding sgRNAs' p-values of the gene.}
\usage{
alphaBeta(pvec)
}
\arguments{
\item{pvec}{A numeric vector of p-values.}
}
\value{
A min value of the kth smallest value based on the beta 
  distribution B(k, n-k+1), where the n is the number of probabiliteis 
  in the vector. This min value is the significance score of the gene.
}
\description{
Code was adapted from R package gscreend.
}
