% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CEDA.R
\name{calculateGeneLFC}
\alias{calculateGeneLFC}
\title{Calculating gene-level log fold ratios}
\usage{
calculateGeneLFC(lfcs, genes)
}
\arguments{
\item{lfcs}{A numeric vector containing log fold change of sgRNAs.}

\item{genes}{A character string containing gene names corresponding to sgRNAs.}
}
\value{
A numeric vector containing log fold ratio of genes.
}
\description{
Log fold ratios of all sgRNAs of a gene are averaged to obtain the 
gene level log fold ratio.
}
