% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CEDA.R
\name{densityPlot}
\alias{densityPlot}
\title{2D density contour plot of gene log2 fold ratios against gene expression levels}
\usage{
densityPlot(data, ...)
}
\arguments{
\item{data}{A data frame from the output of preparePlotData function}

\item{...}{Other graphical parameters}
}
\value{
No return value
}
\description{
This function generates a scatter plot with 2D density contour of log2 fold ratios of sgRNAs
against the corresponding gene expression levels.
}
