% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CEDA.R
\name{EMFit}
\alias{EMFit}
\title{Fitting multi-component normal mixture models by R package mixtools}
\usage{
EMFit(x, k0, mean_constr, sd_constr, npara, d0)
}
\arguments{
\item{x}{A numeric vector}

\item{k0}{Number of components in the normal mixture model}

\item{mean_constr}{A constrain on means of components}

\item{sd_constr}{A constrain on standard deviations of components}

\item{npara}{Number of parameters}

\item{d0}{Number of times for fitting mixture model using different 
starting values}
}
\value{
Normal mixture model fit and BIC value of the log-likelihood
}
\description{
The function normalmixEM in R package mixtools is employed for 
fitting multi-component normal mixture models.
}
