% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mda231.R
\docType{data}
\name{mda231}
\alias{mda231}
\title{CRISPR screen data of cell line MDA-MB-231.}
\format{
A data frame with a list of two elements:
\describe{
  \item{sgRNA}{Raw Read counts of sgRNAs}
  \item{negene}{A list of non-essential genes}
}
}
\usage{
mda231
}
\description{
A dataset containing the expression data of sgRNAs
in a CRISPR screen experiment of cell line MDA-MB-231.
}
\keyword{datasets}
