% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CEDA.R
\name{ridgePlot}
\alias{ridgePlot}
\title{Density ridgeline plot of gene expression levels for different FDR groups.}
\usage{
ridgePlot(data, ...)
}
\arguments{
\item{data}{A data frame from the output of preparePlotData function}

\item{...}{Other graphical parameters}
}
\value{
No return value
}
\description{
This function generates a density ridgeline plot of gene expression levels for different FDR groups.
}
