% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CEDA.R
\name{scatterPlot}
\alias{scatterPlot}
\title{Scatter plot of log2 fold ratios against gene expression levels}
\usage{
scatterPlot(data, fdr, ...)
}
\arguments{
\item{data}{A numeric matrix from the output of normalMM function}

\item{fdr}{A level of false discovery rate}

\item{...}{Other graphical parameters}
}
\value{
No return value
}
\description{
This function generates a scatter plot of log2 fold ratios of sgRNAs
against the corresponding gene expression levels.
}
